Internet fixes
Updated internet file structure
